+++ package.json
@@ -13,6 +13,7 @@
     "react-awesome-styled-grid": "^3.0.4",
     "react-color": "^2.18.0",
     "react-dom": "^16.12.0",
+    "react-grid-layout": "^0.18.2",
     "react-icons": "^3.9.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
