+++ src/components/Dashboard/Charts.js
@@ -3,9 +3,35 @@ import styled from 'styled-components';
 
 import ChartTemplate from './ChartTemplate';
 
+import { Responsive, WidthProvider } from 'react-grid-layout';
+
+const responsiveLayout = WidthProvider(Responsive);
+
+const ResponsiveStyled = styled(responsiveLayout)`
+  width: 100vw;
+`;
+
+const DndBlock = styled.div`
+  height: 400px;
+  width: ${props => props.w - (props.colCount + 1) * 5 + 'px'};
+  background: ${props => props.bg || 'white'};
+  color: ${props => props.color || 'black'};
+
+  box-shadow: 0 0 20px #ddd;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+
+  cursor: pointer;
+
+  transition-duration: 0.1s;
+`;
+
 const ChartContainer = styled.div`
-  min-width: 250px;
-  height: 40vh;
+  /* min-width: 250px;
+  height: 40vh; */
+  position: relative;
   min-height: 300px;
   background: white;
   margin: 2rem;
@@ -27,7 +53,43 @@ const ChartTitle = styled.h4`
 export default class Charts extends Component {
   constructor(props) {
     super(props);
+
+    this.state = {
+      breakPoints: {
+        lg: 1200,
+        md: 996,
+        sm: 768,
+        xs: 480,
+        xxs: 0
+      },
+      cols: { lg: 6, md: 4, sm: 3, xs: 2, xxs: 1 },
+      colCount: null
+    };
   }
+
+  componentDidMount = () => {
+    this.setState({ colCount: this.colCount() });
+    window.addEventListener('resize', this.onBreakPointChange);
+  };
+
+  componentWillUnmount() {
+    window.removeEventListener('resize', this.onBreakPointChange);
+  }
+
+  onBreakPointChange = e => {
+    this.setState({ colCount: null });
+  };
+
+  colCount = () => {
+    const { breakPoints, cols } = this.state;
+    const winW = window.innerWidth;
+
+    for (let point in breakPoints) {
+      if (winW > breakPoints[point]) {
+        return cols[point];
+      }
+    }
+  };
 
   render() {
     let { allData, catKey, catVal, antiKey, timespan } = this.props;
@@ -39,29 +101,76 @@ export default class Charts extends Component {
       )
     );
 
+    const { breakPoints, cols, colCount } = this.state;
+
+    let row = 0;
+
+    const colSize = window.innerWidth / colCount;
+
     return (
       <>
-        {displayedData.map((data, nth) => {
-          // format title
-          const titleKey =
-            antiKey[0].toUpperCase() + antiKey.slice(1).slice(0, -1);
+        <ResponsiveStyled
+          className="gridLayout"
+          breakpoints={breakPoints}
+          cols={cols}
+          rowHeight={400}
+          onBreakpointChange={this.onBreakPointChange}
+          colCount={colCount}
+        >
+          {displayedData.map((data, nth) => {
+            if (nth >= colCount && nth % colCount === 0) row++;
 
-          const titleVal = data[0][antiKey.slice(0, -1)].name;
+            let pos = nth < colCount ? nth : nth % colCount;
 
-          //measure title
-          const flex = titleVal.length > 30 ? 'auto' : '1';
-          const cutYear = titleVal.length > 30 ? false : true;
+            // format title
+            const titleKey =
+              antiKey[0].toUpperCase() + antiKey.slice(1).slice(0, -1);
 
-          return (
-            <ChartContainer key={nth} className="ChartContainer" flex={flex}>
-              <ChartTitle>
-                {titleKey}: {titleVal}
-              </ChartTitle>
-              <ChartTemplate data={data} catVal={catVal} cutYear={cutYear} />
-            </ChartContainer>
-          );
-        })}
+            const titleVal = data[0][antiKey.slice(0, -1)].name;
+
+            //measure title
+            const flex = titleVal.length > 30 ? 'auto' : '1';
+            const cutYear = titleVal.length > 30 ? false : true;
+
+            return (
+              <DndBlock
+                className="DndBlock"
+                data-grid={{ x: pos, y: row, w: 1, h: 1 }}
+                w={colSize}
+                colCount={colCount}
+                key={nth}
+                col={pos}
+                row={row}
+              >
+                <ChartTitle>
+                  {titleKey}: {titleVal}
+                </ChartTitle>
+                <ChartTemplate data={data} catVal={catVal} cutYear={cutYear} />
+              </DndBlock>
+            );
+          })}
+        </ResponsiveStyled>
       </>
     );
   }
 }
+
+/* class DndBlock extends Component {
+  constructor(props) {
+    super(props);
+
+    console.log('block props', this.props.gridProps);
+  }
+
+  render() {
+    const { className, dataGrid, key, w, gridProps, children } = this.props;
+    return (
+      <DndBlockStyled
+        className={className}
+        data-grid={dataGrid}
+        w={w}
+        colCount={gridProps.colCount}
+      ></DndBlockStyled>
+    );
+  }
+} */
